Test for Linear::evaluate_bound
